Feature: removed handshake request.
The handshake request was unnecessary because it was used to get the
XSRF token, this token is however also put on the `cookies` by the
back-end. So getting the token via a request is wasteful.

Now the XSRT token is retrieved from the `document.cookie`, and is
no longer stored in the Redux store.